backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
pdfminer.six
python-docx
nltk
scikit-learn
numpy
pandas
